refactor(heatmap): extract args.rs with table-driven parse_args (#216, #217)
Move parse_args, apply_boolean_flag, extract_value, parse_since_value, and
print_help from mod.rs to a new args.rs module. Also relocate all 31
test_parse_args_* tests and the insights flag parsing tests.

Replace the 10 repeated extract_value+if-let blocks in parse_args with a
table-driven dispatch: VALUE_FLAG_TABLE (const slice of ValueFlagSpec) +
a single loop that calls apply_value_flag(). The old VALUE_FLAGS const is
eliminated. apply_value_flag uses an exhaustive match (no _ catch-all),
preserving exact validation logic and error messages throughout.
